Fix class index out of bounds issue
This error was being caused by the outdated usage of the number
of classes in the `transform_train` function. Another changes
have been made to garantee the number of classes on the training
is consistent while reading the new_label file. The class should
still exists in the classification vector even though the image
is not available on the training set.