Add missing whale landmarks and bounding box
The whale on the 076e1f97f.jpg image is not listed on the
landmark files nor in the train.csv file provided by the author.
However, it is part of our hard example testing.

Therefore, its landmarks and bounding boxes have been included
manually. The left, notch, right and down keypoints were used from
the keypoints.csv file, which was used to train the landmark
prediction network. On the other hand, the bounding box was set
as the minimal box that includes all four keypoints.